#121: refactored LALR and CLR DFA construction into parse package
